--- lite/api-modules/ant/ant.profile
@@ -1,3 +1,0 @@
-ENABLE_MODULE_IOTJS_BASIC_MODULES
-ENABLE_MODULE_IOTJS_CORE_MODULES
-ENABLE_MODULE_ANT
