+++ lite/api-modules/ant/js/ant.js
@@ -1,5 +1,8 @@
 var console = require('console');
 var http = require('http');
+
+var CompanionAPI = require('antcompanion');
+
 var RESULT_SUCCESS = 'Success';
 var RESULT_FAILED = 'Failed';
 
@@ -297,157 +300,6 @@ Element.prototype.link = function (destElement) {
 
 /** Stream API end **/
 
-/** ML API start **/
-function MLAPI() { }
-
-MLAPI.createMLElement = function (model_name, input_shape, input_type, output_shape, output_type) {
-  if (model_name.indexOf(" ") >= 0) {
-    console.error("Invalid model_name! " + model_name);
-    return undefined;
-  }
-  if (input_shape.indexOf(" ") >= 0) {
-    console.error("Invalid input_shape! " + input_shape);
-    return undefined;
-  }
-  if (input_type.indexOf(" ") >= 0) {
-    console.error("Invalid input_type! " + input_type);
-    return undefined;
-  }
-  if (output_shape.indexOf(" ") >= 0) {
-    console.error("Invalid output_shape! " + output_shape);
-    return undefined;
-  }
-  if (output_type.indexOf(" ") >= 0) {
-    console.error("Invalid output_type! " + output_type);
-    return undefined;
-  }
-  if (!StreamAPI.isInitialized()) {
-    console.error("ERROR: Stream API is not initialized");
-    return;
-  }
-
-  function shape_array_to_str(shape_array) {
-    var shape_str = "";
-    for (var i in shape_array) {
-      if (i == 0) {
-        shape_str = shape_str + shape_array[i];
-      } else {
-        shape_str = shape_str + ":" + shape_array[i];
-      }
-    }
-    return shape_str;
-  }
-
-  var input_shape_str = shape_array_to_str(input_shape);
-  var output_shape_str = shape_array_to_str(output_shape);
-
-  var tensor_filter = StreamAPI.createElement("tensor_filter");
-  tensor_filter.setProperty("framework", "python3");
-  tensor_filter.setProperty("model", "./ml/tvm_nnstreamer.py");
-  tensor_filter.setProperty("input", input_shape_str);
-  tensor_filter.setProperty("inputtype", input_type);
-  tensor_filter.setProperty("output", output_shape_str);
-  tensor_filter.setProperty("outputtype", output_type);
-  var custom = model_name + " " + input_shape_str + " " + input_type
-    + " " + output_shape_str + " " + output_type;
-  tensor_filter.setProperty("custom", custom);
-  return tensor_filter;
-};
-
-MLAPI.getMaxOfBuffer = function (buffer, type) {
-  return native.ml_getMaxOfBuffer(buffer, type);
-}
-
-MLAPI.connectCompressionServer = function (ipAddress) {
-  return new CompressionServer(ipAddress);
-};
-
-function CompressionServer(ipAddress) {
-  this.ipAddress = ipAddress;
-}
-
-CompressionServer.prototype.downloadModel = function (modelId) {
-  if (this.ipAddress === undefined)
-    return undefined;
-  // TODO: implement details
-};
-CompressionServer.prototype.downloadKernel = function (kernelId) {
-  if (this.ipAddress === undefined)
-    return undefined;
-  // TODO: implement details
-};
-
-/** NN API end */
-
-/** Companion API start **/
-function CompanionAPI() { }
-CompanionAPI._mCompanionHost = undefined;
-CompanionAPI._mCompanionPort = undefined;
-CompanionAPI._mCompanionPath = undefined;
-CompanionAPI._mHandlers = [];
-
-CompanionAPI._setCompanionAddress = function (
-  companionHost, companionPort, companionPath) {
-  this._mCompanionHost = companionHost;
-  this._mCompanionPort = companionPort;
-  this._mCompanionPath = companionPath;
-  return true;
-};
-
-CompanionAPI._onReceiveMessageFromCompanion = function (message) {
-  for (var i in this._mHandlers) {
-    this._mHandlers[i](message);
-  }
-};
-
-CompanionAPI.sendMessage = function (message) {
-  if (this._mCompanionPath === undefined) {
-    console.log('Error: failed to send message due to no companion address');
-    return false;
-  }
-
-  // IoT.js ----
-  var options = {
-    method: 'POST',
-    host: this._mCompanionHost,
-    port: this._mCompanionPort,
-    path: this._mCompanionPath,
-    headers: { 'Content-Length': message.length },
-  };
-  var client_request = http.request(options);
-  client_request.write(message);
-  client_request.end();
-  // node.js ----
-  // var options = {
-  //   method: 'POST',
-  //   uri: 'http://' + this._mCompanionHost + ':' + this._mCompanionPort + '/' + this._mCompanionPath,
-  //   body: message
-  // };
-  // request.post(options, function (err, httpResponse, body) {
-  // });
-  return true;
-};
-
-CompanionAPI.registerOnReceiveMessage = function (handler) {
-  this._mHandlers.push(handler);
-};
-CompanionAPI.unregisterOnReceiveMessage = function (handler) {
-  if (handler === undefined) return false;
-  for (var i in this._mHandlers) {
-    if (this._mHandlers[i] === handler) {
-      this._mHandlers.splice(i, 1);
-      return true;
-    }
-  }
-  return false;
-};
-
-CompanionAPI.getMyIPAddress = function (interfaceName) {
-  var antcompanion = require("antcompanion");
-  return antcompanion.getMyIPAddress(interfaceName);
-};
-/** Companion API end **/
-
 /** Remote UI API start **/
 function RemoteUIAPI() {
 }
@@ -531,8 +383,8 @@ function ANT() { }
 
 ANT.prototype.runtime = RuntimeAPI;
 ANT.prototype.stream = StreamAPI;
-ANT.prototype.ml = MLAPI;
-ANT.prototype.companion = CompanionAPI;
+ANT.prototype.ml = require('antml');
+ANT.prototype.companion = require('antcompanion');
 ANT.prototype.resource = ResourceAPI;
 ANT.prototype.remoteui = RemoteUIAPI;
