+++ lite/api-modules/ant/native/internal/ant_native_stream.c
@@ -1,4 +1,5 @@
 #include "ant_native_stream.h"
+#include "../../../common/native/ant_common.h"
 
 #include <gio/gio.h>
 #include <glib.h>
