+++ lite/api-modules/antcompanion/native/internal/ant_companion_native_internal.c
@@ -1,4 +1,5 @@
 #include "ant_companion_native_internal.h"
+#include "../../../common/native/ant_common.h"
 
 #include <arpa/inet.h>
 #include <net/if.h>
